requirements.txt
# requirements.txt - Updated for Supabase

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & Supabase
supabase==2.3.4
postgrest==0.13.0
gotrue==2.3.0
realtime==1.0.4
storage3==0.7.4
supafunc==0.5.1

# PostgreSQL with async support
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0

# Vector database support
pgvector==0.2.3
faiss-cpu==1.7.4

# ML Models
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
sentence-transformers==2.2.2
faster-whisper==0.10.0
easyocr==1.7.0
open-clip-torch==2.23.0
opencv-python==4.8.1.78
spacy==3.7.2

# Data processing
numpy==1.24.4
pandas==2.1.4
Pillow==10.1.0
scikit-learn==1.3.2

# Redis for caching
redis==5.0.1
hiredis==2.2.3

# Auth and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.26.0